Got init data working. Need to work on how load data goes for players because right now them loading data doesn't make sense as it doesn't really do anything.
We should probably just make it so that PlayerManager is an IDataPersistence. That's because we can have it store an array of all the players and then when the game loads/saves is when it can init players and load their data. Then it can also save their data individually.