Saving and loading correctly now!
TODO: When we go through character select screen and start the game we enter a "level" while in a level we may be going through different unity scenes with the same players, but at any point if we leave a level, after displaying player scores and other such things, when returning to the menu, we must delete all player data, especially the device values.

Next step: Create a scene for showing the scores of each player at the end of a level, but also ensure we delete all player data (NOT UNLOCKED CHARACTERS)  when returning to any scene other than the next in a level sequence.

the game manager should probably have a line of code to indicate when we lave a level.